+++ pyaem/pyaem.py
@@ -63,6 +63,10 @@ class PyAem(object):
         return self.content_repo.delete_agent(agent_name, run_mode, **kwargs)
 
 
+    def set_property(self, path, property_name, property_value, **kwargs):
+        return self.content_repo.set_property(path, property_name, property_value, **kwargs)
+
+
     # package manager methods