+++ test/test_contentrepo.py
@@ -355,6 +355,29 @@ class TestContentRepo(unittest.TestCase):
             debug=True)
 
 
+    def test_set_property(self):
+
+        _self = self
+        class SetPropertyHandlerMatcher(HandlersMatcher):
+            def __eq__(self, handlers):
+
+                response = None
+                result = handlers[200](response)
+                _self.assertEquals(result.is_success(), True)
+                _self.assertEquals(result.message, 'Set property sling:target=/welcome.html on path /content/mysite')
+                _self.assertEquals(result.response, response)
+
+                return super(SetPropertyHandlerMatcher, self).__eq__(handlers)
+
+        self.content_repo.set_property('/content/mysite', 'sling:target', '/welcome.html', foo='bar')
+        bag.request.assert_called_once_with(
+            'post',
+            'http://localhost:4502/content/mysite',
+            {'sling:target': '/welcome.html', 'foo': 'bar'},
+            SetPropertyHandlerMatcher([200, 405]),
+            debug=True)
+
+
 if __name__ == '__main__':
     unittest.main()