+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add pyaemresult#debug
 * Add packagemanagerservicejsp#is_package_uploaded
 * Add packagemanagerservicejsp#is_package_installed
+* Add content_repo#set_property
 
 ### 0.9.1
 * Add file_path arg to webconsole#install_bundle to allow custom bundle file location